requirements.txt
# Telegram Bot
python-telegram-bot>=21.7

# Google APIs (Sheets, Drive)
gspread>=6.1.2
google-auth>=2.36.0
google-api-python-client>=2.149.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0

# PDF Processing (requires: poppler for pdf2image, tesseract for pytesseract)
pdfplumber>=0.11.4
pytesseract>=0.3.10
Pillow>=11.0.0
pdf2image>=1.17.0

# Word document support
python-docx>=1.1.2

# Gemini API
google-generativeai>=0.8.0
